Refactor assignment with only tax-calculator
There have been two versions of the interview assessment floating
around, so this reduces it to a single, predictable ask. The autonomous
car assignment has been removed in favour of just the tax-calculator.

Included is an update of the instructions that we can reference in
future instead of relying on communication through the emails. It is
written in Markdown and coverted to html on build, which you can find at
the root directory.

Bonuses:
* python bump to version 10
* 2022 tax year + brackets added
* Test cases added for 2022 tqx year

Add values to testable cases

Adjust instructions again